+++ ctr-slc6/64bit/driver/ctrdrvr-vme.c
@@ -2352,7 +2352,7 @@ int rcnt, wcnt;           /* Readable, Writable byte counts at arg address */
       break;
 
       case CtrDrvrSET_UTC:                /* Set Universal Coordinated Time for next PPS tick */
-	 if (lap) return SetTime(mcon,lav);
+	 if (lap) return SetTime(mcon,(unsigned int) lav);
       break;
 
       case CtrDrvrGET_UTC:                /* Latch and read the current UTC time */
